requirements.txt
# Core dependencies
requests>=2.31.0,<3.0.0
cryptography>=41.0.0,<42.0.0
python-dotenv>=1.0.0,<2.0.0

# Date and time handling
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0

# HTTP and networking
urllib3>=2.0.7,<3.0.0
certifi>=2023.7.22,<2024.0
charset-normalizer>=3.3.0,<4.0.0
idna>=3.4,<4.0.0

# Windows-specific (required for GUI)
pywin32>=305; platform_system == "Windows" 